telegram_spacy_bot/requirements.txt
python-telegram-bot>=20.7
spacy>=3.7
pip>=23
setuptools>=68

